Remote control a Shiny app running in a headless browser

This class starts a Shiny app in a new R session, along with a chromote headless browser that can be used to simulate user actions. This provides a full simulation of a Shiny app so that you can test user interactions with a live app.

Methods

Public methods

AppDriver$new()

AppDriver$view()

AppDriver$expect_appshot()

AppDriver$expect_text()

AppDriver$expect_html()

AppDriver$expect_script()

AppDriver$expect_download()

AppDriver$get_values()

AppDriver$get_names()

AppDriver$set_inputs()

AppDriver$click()

AppDriver$upload_file()

AppDriver$wait_for_condition()

AppDriver$wait_for_idle()

AppDriver$wait_for_value()

AppDriver$execute_script()

AppDriver$execute_script_callback()

AppDriver$get_path()

AppDriver$get_url()

AppDriver$get_window_size()

AppDriver$set_window_size()

AppDriver$get_chromote_session()

AppDriver$get_debug_log()

AppDriver$enable_debug_log_messages()

AppDriver$get_event_log()

AppDriver$log_event()

AppDriver$stop()

AppDriver$clone()

Method new()

Initialize the AppDriver

Usage

AppDriver$new(
  path = testthat::test_path("../../"),
  ...,
  load_timeout = NULL,
  screenshot_args = NULL,
  check_names = TRUE,
  name = NULL,
  variant = getOption("shinytest2.variant", platform_variant()),
  debug = c("none", "all", debug_types()),
  view = FALSE,
  seed = NULL,
  clean_logs = TRUE,
  shiny_args = list(),
  render_args = NULL,
  options = list()
)

Arguments

path

Path to a directory containing a Shiny app, i.e. a single app.R file or a server.R-ui.R pair.

...

Must be empty. Allows for parameter expansion.

load_timeout

How long to wait for the app to load, in ms. This includes the time to start R. Defaults to 5s when running locally and 10s when running on CI. Maximum value is 10s.

screenshot_args

Default set of arguments to pass in to chromote::ChromoteSession's $screenshot() method when taking screnshots within $expect_appshot(). To disable screenshots by default, set to FALSE.

check_names

Check if widget names are unique?

name

Prefix name to use when saving testthat snapshot files

variant

If not-NULL, results will be saved in _snaps/variant/test.md, so variant` must be a single string of alphanumeric characters suitable for use as a directory name.

You can variants to deal with cases where the snapshot output varies and you want to capture and test the variations. Common use cases include variations for operating system, R version, or version of key dependency.

debug

Start the app in debugging mode? In debugging mode debug messages are printed to the console. See debug_types() for more information.

view

Opens the Chromote Session in an interactive browser tab once initialization.

seed

An optional random seed to use before starting the application. For apps that use R's random number generator, this can make their behavior repeatable.

clean_logs

Whether to remove the stdout and stderr logs when the Shiny process object is garbage collected.

shiny_args

A list of options to pass to shiny::runApp().

render_args

Passed to rmarkdown::run() for interactive .Rmds.

options

A list of base::options() to set in the driver's child process.

Method view()

Calls $view() on the Chromote Session object

Usage

AppDriver$view()

Method expect_appshot()

Take and appshot of the Shiny application

Appshot: Shiny Application Snapshot

An appshot currently consists of two snapshot files:

A screenshot of the Shiny application using the $screenshot() function of the chromote::ChromoteSession

A JSON snapshot of all Shiny component values

Usage

AppDriver$expect_appshot(
  ...,
  items = NULL,
  screenshot = NULL,
  name = NULL,
  cran = FALSE
)

Arguments

...

Must be empty. Allows for parameter expansion.

items

Components to only be included in the snapshot. If supplied, can contain inputs, output, and export. Each value of items can either be TRUE (for all values) or a character list of names to use.

screenshot

If the value is NULL, then the initalization value of screenshot_args will be used. If this value is NULL, then screenshot will be set to the result of !is.null(items).

The final value can either be:

TRUE: A screenshot of the whole page will be taken with no delay

FALSE: No screenshot will be taken

A named list of arguments: Arguments passed directly to chromote::ChromoteSession's $screenshot() method. The selector and delay will default to "html" and 0 respectively.

name

The prefix name to be used for the snapshot. By default, this uses the name supplied to app on initialization.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method expect_text()

Expect snapshot of UI text

$expect_text() will extract the text value of all matching elements via TAG.textContent and store them in a snapshot file. This method is more robust to internal package change as only the text values will be maintained.

When possible, use $expect_text() over $expect_html() to allow package authors room to alter their HTML structures. The resulting array of TAG.textContent values found will be stored in a snapshot file.

Usage

AppDriver$expect_text(selector, ..., cran = FALSE)

Arguments

selector

A DOM CSS selector to be passed into document.querySelectorAll()

...

Must be empty. Allows for parameter expansion.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method expect_html()

Expect snapshot of UI HTML

$expect_html() will extract the full DOM structures of each matching element and store them in a snapshot file.

Usage

AppDriver$expect_html(selector, ..., outer_html = FALSE, cran = FALSE)

Arguments

selector

A DOM selector to be passed into document.querySelectorAll()

...

Must be empty. Allows for parameter expansion.

outer_html

If TRUE, the full DOM structure will be returned (TAG.outerHTML). If FALSE, the full DOM structure of the child elements will be returned (TAG.innerHTML).

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method expect_script()

Expect snapshot of JS script output

This is a building block function that should be called by other functions. For example, $expect_text() and $expect_html() are thin wrappers around this function.

Usage

AppDriver$expect_script(
  script,
  arguments = list(),
  ...,
  timeout = 15 * 1000,
  pre_snapshot = NULL,
  cran = FALSE
)

Arguments

script

A string containing the JS script to be executed.

arguments

A list of unnamed arguments to send to the script.

...

Must be empty. Allows for parameter expansion.

timeout

Amount of time to wait before giving up (milliseconds).

pre_snapshot

A function to be called on the result of the script before taking the snapshot. $expect_html() and $expect_text() both use unlist().

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method expect_download()

Expect a download file action.

Given a shiny::downloadButton()/shiny::downloadLink() id, the corresponding file will be downloaded and saved as a snapshot file.

Usage

AppDriver$expect_download(id, ..., name = NULL, cran = FALSE)

Arguments

id

Output id of shiny::downloadButton()/shiny::downloadLink()

...

Must be empty. Allows for parameter expansion.

name

File name to save file to. The default, NULL, generates an ascending sequence of names: 001.download, 002.download, etc.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method get_values()

Returns a named list of all inputs, outputs, and export values.

Usage

AppDriver$get_values(input = TRUE, output = TRUE, export = TRUE)

Arguments

input, output, export

Either TRUE to return all input/output/exported values, or a character vector of specific controls.

Method get_names()

Returns a set of names for of all inputs, outputs, and exports that Shiny is currently aware of.

Usage

AppDriver$get_names(input = TRUE, output = TRUE, export = TRUE)

Arguments

input, output, export

Either TRUE to return all input/output/exported values, or FALSE to return no names.

Method set_inputs()

Set input values.

Usage

AppDriver$set_inputs(
  ...,
  wait_ = TRUE,
  values_ = TRUE,
  timeout_ = 3 * 1000,
  allow_input_no_binding_ = FALSE,
  priority_ = c("input", "event")
)

Arguments

...

Name-value pairs, component_name_1 = value_1, component_name_2 = value_2 etc. Input with name component_name_1 will be assigned value value_1.

wait_

Wait until all reactive updates have completed?

values_

If TRUE, will return final updated values of inputs.

timeout_

Amount of time to wait before giving up (milliseconds).

allow_input_no_binding_

When setting the value of an input, allow it to set the value of an input even if that input does not have an input binding.

priority_

Sets the event priority. For expert use only: see https://shiny.rstudio.com/articles/communicating-with-js.html#values-vs-events for details.

Returns

Returns updated values, invisibly.

Method click()

Find a Shiny binding and click it using the DOM method TAG.click()

Usage

AppDriver$click(id, iotype = c("auto", "input", "output"))

Arguments

id

The HTML ID of the element to click

iotype

Type of the Shiny component. shinytest2 is able to find the component by their name, so this is only needed if you use the same name for an input and output component.

Method upload_file()

Uploads a file to a file input.

Usage

AppDriver$upload_file(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3 * 1000)

Arguments

...

Name-path pair, e.g. component_name = file_path. The file located at file_path will be uploaded to file input with name component_name.

wait_

Wait until all reactive updates have completed?

values_

If TRUE, will return final updated values of download control. Otherwise, the return value will be NULL.

timeout_

Amount of time to wait before giving up (milliseconds).

Method wait_for_condition()

Wait for a JavaScript expression to be true

Waits until a JavaScript expression evaluates to true or the timeout is exceeded.

Usage

AppDriver$wait_for_condition(expr, timeout = 3 * 1000, interval = 100)

Arguments

expr

A string containing JavaScript code. Will wait until the condition returns true.

timeout

How often to check for the condition, in ms.

interval

How often to check for the condition, in ms.

Returns

TRUE if expression evaluates to true without error, before timeout. Otherwise returns FALSE.

Method wait_for_idle()

Wait for Shiny to not be busy

Waits until Shiny is not busy, i.e. the reactive graph has finished updating. This is useful, for example, if you've resized the window with $set_window_size() and want to make sure all plot redrawing is complete before take a screenshot.

While this function may return true, Shiny may not have fully stablized. It is best to use TODO-barret-implement wait for stable

Usage

AppDriver$wait_for_idle(timeout = 3 * 1000, interval = 100)

Arguments

timeout

How often to check for the condition, in ms.

interval

How often to check for the condition, in ms.

Returns

TRUE if done before before timeout; NA otherwise.

Method wait_for_value()

Wait for a new Shiny value

Waits until the input or output with name name is not one of ignored values, or the timeout is reached.

This function can be useful in helping determine if an application has initialized or finished processing a complex reactive situation.

Usage

AppDriver$wait_for_value(
  id,
  ignore = list(NULL, ""),
  iotype = c("input", "output", "export"),
  timeout = 10 * 1000,
  interval = 400
)

Arguments

id

Name of the Shiny binding

ignore

List of possible values to ignore when checking for updates.

iotype

Type of the Shiny component. shinytest2 is able to find the component by their name, so this is only needed if you use the same name for an input and output component.

timeout

How often to check for the condition, in ms.

interval

How often to check for the condition, in ms.

Returns

Newly found value

Method execute_script()

Execute JavaScript code in the browser.

This function will block the local R session until the code has finished executing its tick in the browser. If a Promise is returned from the script, $execute_script() will wait for the promise to resolve. To have JavaScript code execute asynchronously, wrap the code in a Promise object and have the script return an atomic value.

Usage

AppDriver$execute_script(script, arguments = list(), ..., timeout = 15 * 1000)

Arguments

script

JS to execute. If a JS Promise is returned, $execute_script() will wait for the promise to resolve before returning.

arguments

A list of unnamed arguments to send to the script.

...

Must be empty. Allows for parameter expansion.

timeout

Amount of time to wait before giving up (milliseconds).

Returns

Result of the script.

Method execute_script_callback()

Execute JavaScript code in the browser with an additional resolve and reject arguments.

This function will block the local R session until one of the last two arguments (resolve and reject) are called.

Usage

AppDriver$execute_script_callback(
  script,
  arguments = list(),
  ...,
  timeout = 15 * 1000
)

Arguments

script

JS to execute. resolve and reject arguments are added to the script call. To return control back to the R session, one of these methods must be called.

arguments

A list of unnamed arguments to send to the script.

...

Must be empty. Allows for parameter expansion.

timeout

Amount of time to wait before giving up (milliseconds).

Returns

Self, invisibly.

Method get_path()

Retrieve the Shiny app path

Usage

AppDriver$get_path()

Returns

If it's a .Rmd file, it will return the full .Rmd path, otherwise it will return the directory containing the file.

Method get_url()

Retrieve the Shiny app URL

Usage

AppDriver$get_url()

Returns

URL

Method get_window_size()

Get current size of the browser window, as list of integer scalars named width and height.

Usage

AppDriver$get_window_size()

Method set_window_size()

Sets size of the browser window.

Usage

AppDriver$set_window_size(width, height)

Arguments

width, height

Height and width of browser, in pixels.

Method get_chromote_session()

Chromote Session object from the chromote package.

Usage

AppDriver$get_chromote_session()

Method get_debug_log()

Query one or more of the debug logs.

Usage

AppDriver$get_debug_log(type = c("all", debug_types()))

Arguments

type

Log type: "all", "shiny_console", "browser", or "shinytest2".

Method enable_debug_log_messages()

Enable/disable debugging messages

Usage

AppDriver$enable_debug_log_messages(enable = TRUE)

Arguments

enable

If TRUE, all Shiny WebSocket messages are recorded. This can be useful for debugging, but can be considered a memory leak in the browser.

Method get_event_log()

Retrieve event log.

Usage

AppDriver$get_event_log()

Method log_event()

Add an event to log.

Usage

AppDriver$log_event(event, ...)

Arguments

event

Event name

...

Addition data to store for event

Method stop()

Stop the Shiny application Stop the app, the terminate external R process that runs the app and the Chromote Session instance.

Usage

AppDriver$stop()

Method clone()

The objects of this class are cloneable with this method.

Usage

AppDriver$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.