Remote control a Shiny app running in a headless browser

This class starts a Shiny app in a new R session, along with a chromote headless browser that can be used to simulate user actions. This provides a full simulation of a Shiny app so that you can test user interactions with a live app.

Methods

Public methods

ShinyDriver2$get_chromote_session()

ShinyDriver2$get_debug_log()

ShinyDriver2$enable_debug_log_messages()

ShinyDriver2$log_event()

ShinyDriver2$get_event_log()

ShinyDriver2$execute_script()

ShinyDriver2$execute_script_callback()

ShinyDriver2$expect_appshot()

ShinyDriver2$expect_js()

ShinyDriver2$expect_text()

ShinyDriver2$expect_html()

ShinyDriver2$get_all_values()

ShinyDriver2$new()

ShinyDriver2$set_inputs()

ShinyDriver2$upload_file()

ShinyDriver2$get_path()

ShinyDriver2$take_screenshot()

ShinyDriver2$expect_download()

ShinyDriver2$stop()

ShinyDriver2$wait_for_condition()

ShinyDriver2$wait_for_idle()

ShinyDriver2$wait_for_value()

ShinyDriver2$get_value()

ShinyDriver2$set_value()

ShinyDriver2$click()

ShinyDriver2$list_widgets()

ShinyDriver2$set_window_size()

ShinyDriver2$get_window_size()

ShinyDriver2$clone()

Method get_chromote_session()

Usage

ShinyDriver2$get_chromote_session()

Method get_debug_log()

Usage

ShinyDriver2$get_debug_log(type = c("all", debug_log_types()))

Method enable_debug_log_messages()

Usage

ShinyDriver2$enable_debug_log_messages(enable = TRUE)

Method log_event()

Usage

ShinyDriver2$log_event(event, ...)

Method get_event_log()

Usage

ShinyDriver2$get_event_log()

Method execute_script()

Usage

ShinyDriver2$execute_script(script, ...)

Method execute_script_callback()

Usage

ShinyDriver2$execute_script_callback(script, ..., timeout_ = 15 * 1000)

Method expect_appshot()

Usage

ShinyDriver2$expect_appshot(
  ...,
  name = NULL,
  items = NULL,
  screenshot = NULL,
  error = FALSE,
  cran = FALSE
)

Arguments

name

Name of a shiny widget.

Method expect_js()

Usage

ShinyDriver2$expect_js(script, ..., post_fn = NULL, cran = FALSE)

Method expect_text()

Usage

ShinyDriver2$expect_text(selector, ..., cran = FALSE)

Method expect_html()

Usage

ShinyDriver2$expect_html(selector, ..., outer_html = FALSE, cran = FALSE)

Method get_all_values()

Usage

ShinyDriver2$get_all_values(input = TRUE, output = TRUE, export = TRUE)

Method new()

Usage

ShinyDriver2$new(
  path = ".",
  ...,
  load_timeout = NULL,
  screenshot = TRUE,
  check_names = TRUE,
  name = NULL,
  variant = getOption("shinytest2.variant", os_name_and_r_version()),
  debug = c("none", "all", debug_log_types()),
  seed = NULL,
  clean_logs = TRUE,
  shiny_args = list(),
  render_args = NULL,
  options = list()
)

Arguments

name

Name of a shiny widget.

Method set_inputs()

Usage

ShinyDriver2$set_inputs(
  ...,
  wait_ = TRUE,
  values_ = TRUE,
  timeout_ = 3000,
  allow_input_no_binding_ = FALSE,
  priority_ = c("input", "event")
)

Arguments

wait_

Wait until all reactive updates have completed?

Method upload_file()

Usage

ShinyDriver2$upload_file(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3000)

Arguments

wait_

Wait until all reactive updates have completed?

Method get_path()

Usage

ShinyDriver2$get_path()

Method take_screenshot()

Usage

ShinyDriver2$take_screenshot(
  filename = NULL,
  ...,
  screenshot_args = list(),
  delay = 0,
  selector = "html",
  wait_ = TRUE
)

Arguments

wait_

Wait until all reactive updates have completed?

Method expect_download()

Usage

ShinyDriver2$expect_download(id, ..., name = NULL, cran = FALSE)

Arguments

name

Name of a shiny widget.

Method stop()

Usage

ShinyDriver2$stop()

Method wait_for_condition()

Usage

ShinyDriver2$wait_for_condition(expr, timeout = 3 * 1000, interval = 100)

Method wait_for_idle()

Usage

ShinyDriver2$wait_for_idle(timeout = 3 * 1000, interval = 100)

Method wait_for_value()

Usage

ShinyDriver2$wait_for_value(
  name,
  ignore = list(NULL, ""),
  iotype = c("input", "output", "export"),
  timeout = 10000,
  check_interval = 400
)

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method get_value()

Usage

ShinyDriver2$get_value(name, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method set_value()

Usage

ShinyDriver2$set_value(name, value, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method click()

Usage

ShinyDriver2$click(name, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method list_widgets()

Usage

ShinyDriver2$list_widgets()

Method set_window_size()

Usage

ShinyDriver2$set_window_size(width, height)

Method get_window_size()

Usage

ShinyDriver2$get_window_size()

Method clone()

The objects of this class are cloneable with this method.

Usage

ShinyDriver2$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.