Expect a downloadable file

Expect a downloadable file

app_expect_download(app, id, ..., name = NULL, cran = FALSE)

Arguments

app
A ShinyDriver2 object.

id
Output id of shiny::downloadButton()/shiny::downloadLink()

...
Must be empty. Allows for parameter expansion.

name
File name to save file to. The default, NULL, generates an ascending sequence of names: 001.download, 002.download, etc.

cran
Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.