Execute a JavaScript script

Given wait_ = TRUE, chromote_execute_script() will block the testing R session until the script has completed the JS execution tick within the headless browser.

chromote_execute_script(
  chromote,
  js_script,
  ...,
  awaitPromise = wait_,
  arguments = list(),
  timeout = 10 * 1000,
  wait_ = TRUE
)

chromote_execute_script_callback(
  chromote,
  js_script,
  ...,
  arguments = list2(),
  timeout = 15 * 1000,
  wait_ = TRUE
)

chromote_wait_for_condition(
  chromote,
  condition_js,
  ...,
  timeout = 15 * 1000,
  interval = 100,
  wait_ = TRUE
)

Arguments

chromote
A ChromoteSession object

js_script
A string containing the script to be evaluated

...
Arguments passed to chromote$Runtime$evaluate, such as wait_

awaitPromise
If TRUE (default), the function will wait for the Promise object returned from js_script. If wait_ == FALSE, awaitPromise will overwritten to FALSE.

arguments
An unnamed list of arguments to be passed into the js_script

timeout
The maximum time (milliseconds) chromote will wait for the js_script to resolved

wait_
Determines if chromote should return a Promise object or wait for the js_script to be resolved

condition_js
A piece of JavaScript code that should eventually evaluate to a truethy value.

interval
How long (milliseconds) Chrome should wait between checking condition_js

Value

TRUE if expression evaluates to TRUE without error, before timeout. Otherwise returns FALSE.

Functions

chromote_execute_script: Executes the supplied Javascript script (js_script) within a function. The function has the window context and access to arguments supplied.

chromote_execute_script_callback: Executes the supplied Javascript script (js_script) within a function. The function has the window context and access to arguments supplied. An extra argument (resolve(val)) is added to the arguments list. If wait_ = TRUE, then chromote_execute_script_callback() will block the main R session until resolve() has been called.

chromote_wait_for_condition: Waits for condition_js to return a truethy value. In this con