Launch test event recorder for a Shiny app

Once a recording is completed, it will create or append a new shinytest2 test to the testthat test_file.

record_test(
  app = ".",
  ...,
  name = NULL,
  seed = NULL,
  load_timeout = NULL,
  shiny_args = list(),
  test_file = "test-shinytest2.R",
  edit_test = rlang::is_interactive() && rstudioapi::isAvailable(),
  allow_input_no_binding = NULL,
  run_test = TRUE
)

Arguments

app

A AppDriver object, or path to a Shiny application.

...

Must be empty. Allows for parameter expansion.

name

Name provided to AppDriver. This value should be unique between all tests within a test file. If it is not unique, different expect methods may overwrite each other.

seed

A random seed to set before running the app. This seed will also be used in the test script.

load_timeout

Maximum time to wait for the Shiny application to load, in milliseconds. If a value is provided, it will be saved in the test script.

shiny_args

A list of options to pass to runApp(). If a value is provided, it will be saved in the test script.

test_file

Base file name of the testthat test file.

edit_test

If TRUE, the test file will be opened in an editor via file.edit() before executing.

allow_input_no_binding

This value controls if events without input bindings are recorded.

If TRUE, events without input bindings are recorded.

If FALSE, events without input bindings are not recorded.

If NULL (default), if an updated input does not have a corresponding input, a modal dialog will be shown asking if unbound input events should be recorded. See AppDriver$set_inputs() for more information.

run_test

If TRUE, test_file will be executed after saving the recording.