Remote control a Shiny app running in a headless browser

This class starts a Shiny app in a new R session, along with a chromote headless browser that can be used to simulate user actions. This provides a full simulation of a Shiny app so that you can test user interactions with a live app.

Startup failure

If the app throws an error during initialization, the AppDriver will will be stored in rlang::last_error()$app. This allows for the "failure to initialize" to be signaled while also allowing for the app to be retrieved after any initialization error has been thrown.

Methods

Public methods

AppDriver$new()

AppDriver$view()

AppDriver$expect_text()

AppDriver$get_text()

AppDriver$expect_html()

AppDriver$get_html()

AppDriver$expect_js()

AppDriver$expect_download()

AppDriver$get_value()

AppDriver$get_values()

AppDriver$expect_values()

AppDriver$screenshot()

AppDriver$expect_screenshot()

AppDriver$set_inputs()

AppDriver$click()

AppDriver$upload_file()

AppDriver$wait_for_js()

AppDriver$wait_for_idle()

AppDriver$wait_for_value()

AppDriver$execute_js()

AppDriver$get_path()

AppDriver$get_url()

AppDriver$get_window_size()

AppDriver$set_window_size()

AppDriver$get_chromote_session()

AppDriver$get_variant()

AppDriver$get_log()

AppDriver$log_message()

AppDriver$stop()

Method new()

Initialize the AppDriver

Usage

AppDriver$new(
  path = testthat::test_path("../../"),
  ...,
  load_timeout = NULL,
  variant = getOption("shinytest2.variant", missing_arg()),
  expect_values_screenshot_args = TRUE,
  screenshot_args = missing_arg(),
  check_names = TRUE,
  name = NULL,
  view = missing_arg(),
  seed = NULL,
  clean_logs = TRUE,
  shiny_args = list(),
  render_args = NULL,
  options = list()
)

Arguments

path

Path to a directory containing a Shiny app, i.e. a single app.R file or a server.R-ui.R pair.

...

Must be empty. Allows for parameter expansion.

load_timeout

How long to wait for the app to load, in ms. This includes the time to start R. Defaults to 10s when running locally and 20s when running on CI.

variant

If not-NULL, results will be saved in _snaps/variant/test.md, so variant` must be a single string of alphanumeric characters suitable for use as a directory name.

You can variants to deal with cases where the snapshot output varies and you want to capture and test the variations. Common use cases include variations for operating system, R version, or version of key dependency. For example usage, see platform_variant().

screenshot_args

Default set of arguments to pass in to chromote::ChromoteSession's $screenshot() method when taking screnshots within $expect_screenshot(). To disable screenshots by default, set to FALSE.

check_names

Check if widget names are unique?

name

Prefix name to use when saving testthat snapshot files

view

Opens the Chromote Session in an interactive browser tab once initialization. Defaults to FALSE.

seed

An optional random seed to use before starting the application. For apps that use R's random number generator, this can make their behavior repeatable.

clean_logs

Whether to remove the stdout and stderr logs when the Shiny process object is garbage collected.

shiny_args

A list of options to pass to shiny::runApp().

render_args

Passed to rmarkdown::run() for interactive .Rmds.

options

A list of base::options() to set in the driver's child process. See shiny::shinyOptions() for inspiration. If shiny.trace is set to TRUE, then all WebSocket traffic will be captured by chromote as to have access to the when the message was received by the browser.

Method view()

Calls $view() on the Chromote Session object

Usage

AppDriver$view()

Method expect_text()

Expect snapshot of UI text

$expect_text() will extract the text value of all matching elements via TAG.textContent and store them in a snapshot file. This method is more robust to internal package change as only the text values will be maintained.

When possible, use $expect_text() over $expect_html() to allow package authors room to alter their HTML structures. The resulting array of TAG.textContent values found will be stored in a snapshot file.

Usage

AppDriver$expect_text(selector, ..., cran = FALSE)

Arguments

selector

A DOM CSS selector to be passed into document.querySelectorAll()

...

Must be empty. Allows for parameter expansion.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method get_text()

Usage

AppDriver$get_text(selector)

Arguments

selector

A DOM CSS selector to be passed into document.querySelectorAll()

Method expect_html()

Expect snapshot of UI HTML

$expect_html() will extract the full DOM structures of each matching element and store them in a snapshot file.

Usage

AppDriver$expect_html(selector, ..., outer_html = FALSE, cran = FALSE)

Arguments

selector

A DOM selector to be passed into document.querySelectorAll()

...

Must be empty. Allows for parameter expansion.

outer_html

If TRUE, the full DOM structure will be returned (TAG.outerHTML). If FALSE, the full DOM structure of the child elements will be returned (TAG.innerHTML).

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method get_html()

Usage

AppDriver$get_html(selector, ..., outer_html = FALSE)

Arguments

selector

A DOM selector to be passed into document.querySelectorAll()

...

Must be empty. Allows for parameter expansion.

outer_html

If TRUE, the full DOM structure will be returned (TAG.outerHTML). If FALSE, the full DOM structure of the child elements will be returned (TAG.innerHTML).

Method expect_js()

Expect snapshot of JS script output

This is a building block function that should be called by other functions. For example, $expect_text() and $expect_html() are thin wrappers around this function.

Usage

AppDriver$expect_js(
  script = missing_arg(),
  ...,
  arguments = list(),
  file = missing_arg(),
  timeout = 15 * 1000,
  pre_snapshot = NULL,
  cran = FALSE
)

Arguments

script

A string containing the JS script to be executed.

...

Must be empty. Allows for parameter expansion.

arguments

A list of unnamed arguments to send to the script.

file

A file containing JavaScript code to be read and used as the script. Only one of script or file can be specified.

timeout

Amount of time to wait before giving up (milliseconds).

pre_snapshot

A function to be called on the result of the script before taking the snapshot. $expect_html() and $expect_text() both use unlist().

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method expect_download()

Expect a download file action.

Given a shiny::downloadButton()/shiny::downloadLink() id, the corresponding file will be downloaded and saved as a snapshot file.

Usage

AppDriver$expect_download(id, ..., name = NULL, cran = FALSE)

Arguments

id

Output id of shiny::downloadButton()/shiny::downloadLink()

...

Must be empty. Allows for parameter expansion.

name

File name to save file to. The default, NULL, generates an ascending sequence of names: 001.download, 002.download, etc.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method get_value()

Returns a named list of all inputs, outputs, and export values.

Usage

AppDriver$get_value(
  ...,
  input = missing_arg(),
  output = missing_arg(),
  export = missing_arg(),
  hash_images = FALSE
)

Arguments

...

Must be empty. Allows for parameter expansion.

input, output, export

One of these variable should contain a single string value. If more than one value is specified or no values are specified, an error will be thrown.

Method get_values()

Returns a named list of all inputs, outputs, and export values.

Usage

AppDriver$get_values(
  ...,
  input = missing_arg(),
  output = missing_arg(),
  export = missing_arg(),
  hash_images = FALSE
)

Arguments

...

Must be empty. Allows for parameter expansion.

input, output, export

Either TRUE to return all input/output/exported values, or a character vector of specific controls.

Method expect_values()

Take and appshot of the Shiny application

Appshot: Shiny Application Snapshot

An appshot currently consists of two snapshot files:

A screenshot of the Shiny application using the $screenshot() function of the chromote::ChromoteSession

A JSON snapshot of all Shiny component values

Usage

AppDriver$expect_values(
  ...,
  input = missing_arg(),
  output = missing_arg(),
  export = missing_arg(),
  screenshot_args = missing_arg(),
  name = NULL,
  cran = FALSE
)

Arguments

...

Must be empty. Allows for parameter expansion.

name

The prefix name to be used for the snapshot. By default, this uses the name supplied to app on initialization.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

items

Components to only be included in the snapshot. If supplied, can contain inputs, output, and export. Each value of items can either be TRUE (for all values) or a character list of names to use.

screenshot

If the value is NULL, then the initalization value of screenshot_args will be used. If this value is NULL, then screenshot will be set to the result of !is.null(items).

The final value can either be:

TRUE: A screenshot of the whole page will be taken with no delay

FALSE: No screenshot will be taken

A named list of arguments: Arguments passed directly to chromote::ChromoteSession's $screenshot() method. The selector and delay will default to "html" and 0 respectively.

Method screenshot()

Usage

AppDriver$screenshot(
  file = NULL,
  ...,
  screenshot_args = missing_arg(),
  delay = missing_arg(),
  selector = missing_arg()
)

Arguments

...

Must be empty. Allows for parameter expansion.

Method expect_screenshot()

Usage

AppDriver$expect_screenshot(
  ...,
  screenshot_args = missing_arg(),
  delay = missing_arg(),
  selector = missing_arg(),
  name = NULL,
  cran = FALSE
)

Arguments

...

Must be empty. Allows for parameter expansion.

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

Method set_inputs()

Set input values.

Usage

AppDriver$set_inputs(
  ...,
  wait_ = TRUE,
  timeout_ = 3 * 1000,
  allow_input_no_binding_ = FALSE,
  priority_ = c("input", "event")
)

Arguments

...

Name-value pairs, component_name_1 = value_1, component_name_2 = value_2 etc. Input with name component_name_1 will be assigned value value_1.

wait_

Wait until all reactive updates have completed?

timeout_

Amount of time to wait before giving up (milliseconds).

allow_input_no_binding_

When setting the value of an input, allow it to set the value of an input even if that input does not have an input binding.

priority_

Sets the event priority. For expert use only: see https://shiny.rstudio.com/articles/communicating-with-js.html#values-vs-events for details.

Returns

Returns updated values, invisibly.

Method click()

Find a Shiny input/output value or DOM CSS selector and click it using the DOM method TAG.click()

Usage

AppDriver$click(
  input = missing_arg(),
  output = missing_arg(),
  selector = missing_arg()
)

Arguments

input, output, selector

A name of an Shiny input/output value or a DOM CSS selector. Only one of these may be used.

Method upload_file()

Uploads a file to a file input.

Usage

AppDriver$upload_file(..., wait_ = TRUE, timeout_ = 3 * 1000)

Arguments

...

Name-path pair, e.g. component_name = file_path. The file located at file_path will be uploaded to file input with name component_name.

wait_

Wait until all reactive updates have completed?

timeout_

Amount of time to wait before giving up (milliseconds).

Method wait_for_js()

Wait for a JavaScript expression to be true

Waits until a JavaScript expression evaluates to true or the timeout is exceeded.

Usage

AppDriver$wait_for_js(script, timeout = 3 * 1000, interval = 100)

Arguments

script

A string containing JavaScript code. This code must eventually return a truethy value or a timeout error will be thrown.

timeout

How long the script has to return a truethy value, in ms.

interval

How often to check for the condition, in ms.

Returns

invisible(self) if expression evaluates to true without error within the timeout. Otherwise an error will be thrown

Method wait_for_idle()

Wait for Shiny to not be busy (idle) for a set amount of time

Waits until Shiny has not been busy for a set duration of time, i.e. no reactivity is updating or has occured. This is useful, for example, when waiting for your application to initialize or if you've resized the window with $set_window_size() and want to make sure all plot redrawing is complete before take a screenshot.

Usage

AppDriver$wait_for_idle(duration = 500, timeout = 30 * 1000)

Arguments

duration

How long Shiny must be idle (in ms) before unblocking the R session.

timeout

How often to check for the condition, in ms.

Returns

invisible(self) if Shiny stablizes within the timeout. Otherwise an error will be thrown

Method wait_for_value()

Wait for a new Shiny value

Waits until input, output, or exported shiny value is not one of ignored values, or the timeout is reached.

Only a single input, output, or export value may be used.

This function can be useful in helping determine if an application has finished processing a complex reactive situation.

Usage

AppDriver$wait_for_value(
  ...,
  input = missing_arg(),
  output = missing_arg(),
  export = missing_arg(),
  ignore = list(NULL, ""),
  timeout = 15 * 1000,
  interval = 400
)

Arguments

...

Must be empty. Allows for parameter expansion.

input, output, export

A name of an input, output, or export value. Only one of these parameters may be used.

ignore

List of possible values to ignore when checking for updates.

timeout

How long we can wait (in ms) before throwing an error.

interval

How often to check for the condition, in ms.

Returns

Newly found value

Method execute_js()

Execute JavaScript code in the browser.

This function will block the local R session until the code has finished executing its tick in the browser. If a Promise is returned from the script, $execute_js() will wait for the promise to resolve. To have JavaScript code execute asynchronously, wrap the code in a Promise object and have the script return an atomic value.

Usage

AppDriver$execute_js(
  script = missing_arg(),
  ...,
  arguments = list(),
  file = missing_arg(),
  timeout = 15 * 1000
)

Arguments

script

JS to execute. If a JS Promise is returned, the R session will block until the promise has been resolved and return the value.

...

Must be empty. Allows for parameter expansion.

arguments

A list of unnamed arguments to send to the script.

file

A (local) file containing JavaScript code to be read and used as the script. Only one of script or file can be specified.

timeout

Amount of time to wait before giving up (milliseconds).

Returns

Result of the script (or file contents)

Method get_path()

Retrieve the Shiny app path

Usage

AppDriver$get_path()

Returns

If it's a .Rmd file, it will return the full .Rmd path, otherwise it will return the directory containing the file.

Method get_url()

Retrieve the Shiny app URL

Usage

AppDriver$get_url()

Returns

URL

Method get_window_size()

Get current size of the browser window, as list of integer scalars named width and height.

Usage

AppDriver$get_window_size()

Method set_window_size()

Sets size of the browser window.

Usage

AppDriver$set_window_size(width, height)

Arguments

width, height

Height and width of browser, in pixels.

Method get_chromote_session()

Chromote Session object from the chromote package.

Usage

AppDriver$get_chromote_session()

Method get_variant()

Get the variant supplied at initialization

Usage

AppDriver$get_variant()

Method get_log()

Query one or more of the debug logs. There are a few standard debug types that may be used:

"shiny_console": Displays the console messages from the Shiny server when $get_log() is called.

"browser": Displays the browser console messages when $get_log() is called.

"shinytest2": Displays the messages saved by the window.shinytest2 object in the browser when $get_log() is called.

"ws_messages": Saves all messages sent by Shiny to the

Usage

AppDriver$get_log()

Returns

A data.frame with the following columns:

workerid: The shiny worker ID found within the browser

timestamp: POSIXct timestamp of the message

location: The location of the message was found. One of three values:

shinytest2: Occurs when $log_message() is called

shiny: Stdin and stdout messages from the Shiny server. Note message() output is sent to stdout.

chromote: Captured by the chromote event handlers. See console API, exception thrown, websocket sent, and websocket received for more details

level:

Examples

\dontrun{

app <- AppDriver$new(system.file("examples/01_hello", package = "shiny"))
app$get_log()
# {shinytest2} R  info  11:15:20.11 Start AppDriver initialization
# {shinytest2} R  info  11:15:20.11 Starting Shiny app
# {shinytest2} R  info  11:15:20.99 Creating new chromote session
# {shinytest2} R  info  11:15:21.14 Navigating to Shiny app
# {shinytest2} R  info  11:15:21.27 Injecting shiny-tracer.js
# {chromote}   JS info  11:15:21.28 shinytest2; jQuery not found
# {chromote}   JS info  11:15:21.28 shinytest2; Loaded
# {shinytest2} R  info  11:15:21.28 Waiting until Shiny app starts
# {chromote}   JS info  11:15:21.35 shinytest2; jQuery found
# {chromote}   JS info  11:15:21.35 shinytest2; Waiting for shiny session to connect
# {chromote}   JS info  11:15:21.57 shinytest2; Connected
# {chromote}   JS info  11:15:21.57 shinytest2; Ready
# {chromote}   JS info  11:15:21.65 shinytest2; shiny:busy
# {shinytest2} R  info  11:15:21.65 Shiny app started
# {chromote}   JS info  11:15:21.88 shinytest2; shiny:idle
# {chromote}   JS info  11:15:21.88 shinytest2; shiny:value distPlot
# {shiny}      R  error ----------- Loading required package: shiny
# {shiny}      R  error ----------- Running application in test mode.
# {shiny}      R  error -----------
# {shiny}      R  error ----------- Listening on http://127.0.0.1:42558

# To capture all websocket traffic, set `options = list(shiny.trace = TRUE)`
app <- AppDriver$new(
  system.file("examples/01_hello", package = "shiny"),
  options = list(shiny.trace = TRUE)
)
app$get_log() # (long output lines have been truncated)
# {shinytest2} R  info      11:09:57.43 Start AppDriver initialization
# {shinytest2} R  info      11:09:57.43 Starting Shiny app
# {shinytest2} R  info      11:09:58.27 Creating new chromote session
# {shinytest2} R  info      11:09:58.40 Navigating to Shiny app
# {shinytest2} R  info      11:09:58.53 Injecting shiny-tracer.js
# {chromote}   JS info      11:09:58.53 shinytest2; jQuery not found
# {chromote}   JS info      11:09:58.53 shinytest2; Loaded
# {shinytest2} R  info      11:09:58.54 Waiting until Shiny app starts
# {chromote}   JS info      11:09:58.61 shinytest2; jQuery found
# {chromote}   JS info      11:09:58.61 shinytest2; Waiting for shiny session to connect
# {chromote}   JS websocket 11:09:58.73 send {"method":"init","data":{"bins":30,|truncated
# {chromote}   JS websocket 11:09:58.78 recv {"config":{"workerId":"","sessionId":|truncated
# {chromote}   JS info      11:09:58.78 shinytest2; Connected
# {chromote}   JS info      11:09:58.78 shinytest2; Ready
# {chromote}   JS websocket 11:09:58.85 recv {"custom":{"showcase-src":{"srcref":|truncated
# {chromote}   JS websocket 11:09:58.85 recv {"busy":"busy"}
# {chromote}   JS info      11:09:58.85 shinytest2; shiny:busy
# {chromote}   JS websocket 11:09:58.86 recv {"custom":{"showcase-src":{"srcref":|truncated
# {chromote}   JS websocket 11:09:58.86 recv {"recalculating":{"name":"distPlot",|truncated
# {shinytest2} R  info      11:09:58.87 Shiny app started
# {shinytest2} R  info      11:09:59.07 Setting inputs: 'bins'
# {chromote}   JS websocket 11:09:59.08 recv {"recalculating":{"name":"distPlot",|truncated
# {chromote}   JS websocket 11:09:59.08 recv {"busy":"idle"}
# {chromote}   JS info      11:09:59.08 shinytest2; shiny:idle
# {chromote}   JS websocket 11:09:59.08 recv {"errors":{},"values":{"distPlot":{|truncated
# {chromote}   JS info      11:09:59.08 shinytest2; shiny:value distPlot
# {chromote}   JS info      11:09:59.08 shinytest2; inputQueue: adding bins
# {chromote}   JS info      11:09:59.09 shinytest2; inputQueue: flushing bins
# {chromote}   JS websocket 11:09:59.10 send {"method":"update","data":{"bins":20}}
# {chromote}   JS websocket 11:09:59.11 recv {"progress":{"type":"binding",|truncated
# {chromote}   JS websocket 11:09:59.11 recv {"busy":"busy"}
# {chromote}   JS info      11:09:59.11 shinytest2; shiny:busy
# {chromote}   JS websocket 11:09:59.12 recv {"custom":{"showcase-src":{"srcref":|truncated
# {chromote}   JS websocket 11:09:59.14 recv {"recalculating":{"name":"distPlot",|truncated
# {chromote}   JS websocket 11:09:59.18 recv {"recalculating":{"name":"distPlot",|truncated
# {chromote}   JS websocket 11:09:59.19 recv {"busy":"idle"}
# {chromote}   JS info      11:09:59.19 shinytest2; shiny:idle
# {chromote}   JS websocket 11:09:59.21 recv {"errors":{},"values":{"distPlot":{|truncated
# {chromote}   JS info      11:09:59.21 shinytest2; shiny:value distPlot
# {shinytest2} R  info      11:09:59.21 Finished setting inputs. Timedout: FALSE
# {shinytest2} R  info      11:09:59.21 Getting all values
# {shiny}      R  error     ----------- Loading required package: shiny
# {shiny}      R  error     ----------- Running application in test mode.
# {shiny}      R  error     -----------
# {shiny}      R  error     ----------- Listening on http://127.0.0.1:1505
# {shiny}      R  error     ----------- SEND {"config":{"workerId":"","sessionId":|truncated
# {shiny}      R  error     ----------- RECV {"method":"init","data":{"bins":30,|truncated
# {shiny}      R  error     ----------- SEND {"custom":{"showcase-src":{"srcref"|truncated
# {shiny}      R  error     ----------- SEND {"busy":"busy"}
# {shiny}      R  error     ----------- SEND {"custom":{"showcase-src":{"srcref"|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"busy":"idle"}
# {shiny}      R  error     ----------- SEND {"errors":{},"values":{"distPlot":{|truncated
# {shiny}      R  error     ----------- RECV {"method":"update","data":{"bins":20}}
# {shiny}      R  error     ----------- SEND {"progress":{"type":"binding",|truncated
# {shiny}      R  error     ----------- SEND {"busy":"busy"}
# {shiny}      R  error     ----------- SEND {"custom":{"showcase-src":{"srcref":|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"busy":"idle"}
# {shiny}      R  error     ----------- SEND {"errors":{},"values":{"distPlot":{|truncated

}

Method log_message()

Add an message to log.

Usage

AppDriver$log_message(message)

Arguments

message

Single message to store in log

Method stop()

Stop the Shiny application Stop the app, the terminate external R process that runs the app and the Chromote Session instance.

Usage

AppDriver$stop()

Examples

## ------------------------------------------------
## Method `AppDriver$get_log`
## ------------------------------------------------

if (FALSE) {

app <- AppDriver$new(system.file("examples/01_hello", package = "shiny"))
app$get_log()
# {shinytest2} R  info  11:15:20.11 Start AppDriver initialization
# {shinytest2} R  info  11:15:20.11 Starting Shiny app
# {shinytest2} R  info  11:15:20.99 Creating new chromote session
# {shinytest2} R  info  11:15:21.14 Navigating to Shiny app
# {shinytest2} R  info  11:15:21.27 Injecting shiny-tracer.js
# {chromote}   JS info  11:15:21.28 shinytest2; jQuery not found
# {chromote}   JS info  11:15:21.28 shinytest2; Loaded
# {shinytest2} R  info  11:15:21.28 Waiting until Shiny app starts
# {chromote}   JS info  11:15:21.35 shinytest2; jQuery found
# {chromote}   JS info  11:15:21.35 shinytest2; Waiting for shiny session to connect
# {chromote}   JS info  11:15:21.57 shinytest2; Connected
# {chromote}   JS info  11:15:21.57 shinytest2; Ready
# {chromote}   JS info  11:15:21.65 shinytest2; shiny:busy
# {shinytest2} R  info  11:15:21.65 Shiny app started
# {chromote}   JS info  11:15:21.88 shinytest2; shiny:idle
# {chromote}   JS info  11:15:21.88 shinytest2; shiny:value distPlot
# {shiny}      R  error ----------- Loading required package: shiny
# {shiny}      R  error ----------- Running application in test mode.
# {shiny}      R  error -----------
# {shiny}      R  error ----------- Listening on http://127.0.0.1:42558

# To capture all websocket traffic, set `options = list(shiny.trace = TRUE)`
app <- AppDriver$new(
  system.file("examples/01_hello", package = "shiny"),
  options = list(shiny.trace = TRUE)
)
app$get_log() # (long output lines have been truncated)
# {shinytest2} R  info      11:09:57.43 Start AppDriver initialization
# {shinytest2} R  info      11:09:57.43 Starting Shiny app
# {shinytest2} R  info      11:09:58.27 Creating new chromote session
# {shinytest2} R  info      11:09:58.40 Navigating to Shiny app
# {shinytest2} R  info      11:09:58.53 Injecting shiny-tracer.js
# {chromote}   JS info      11:09:58.53 shinytest2; jQuery not found
# {chromote}   JS info      11:09:58.53 shinytest2; Loaded
# {shinytest2} R  info      11:09:58.54 Waiting until Shiny app starts
# {chromote}   JS info      11:09:58.61 shinytest2; jQuery found
# {chromote}   JS info      11:09:58.61 shinytest2; Waiting for shiny session to connect
# {chromote}   JS websocket 11:09:58.73 send {"method":"init","data":{"bins":30,|truncated
# {chromote}   JS websocket 11:09:58.78 recv {"config":{"workerId":"","sessionId":|truncated
# {chromote}   JS info      11:09:58.78 shinytest2; Connected
# {chromote}   JS info      11:09:58.78 shinytest2; Ready
# {chromote}   JS websocket 11:09:58.85 recv {"custom":{"showcase-src":{"srcref":|truncated
# {chromote}   JS websocket 11:09:58.85 recv {"busy":"busy"}
# {chromote}   JS info      11:09:58.85 shinytest2; shiny:busy
# {chromote}   JS websocket 11:09:58.86 recv {"custom":{"showcase-src":{"srcref":|truncated
# {chromote}   JS websocket 11:09:58.86 recv {"recalculating":{"name":"distPlot",|truncated
# {shinytest2} R  info      11:09:58.87 Shiny app started
# {shinytest2} R  info      11:09:59.07 Setting inputs: 'bins'
# {chromote}   JS websocket 11:09:59.08 recv {"recalculating":{"name":"distPlot",|truncated
# {chromote}   JS websocket 11:09:59.08 recv {"busy":"idle"}
# {chromote}   JS info      11:09:59.08 shinytest2; shiny:idle
# {chromote}   JS websocket 11:09:59.08 recv {"errors":{},"values":{"distPlot":{|truncated
# {chromote}   JS info      11:09:59.08 shinytest2; shiny:value distPlot
# {chromote}   JS info      11:09:59.08 shinytest2; inputQueue: adding bins
# {chromote}   JS info      11:09:59.09 shinytest2; inputQueue: flushing bins
# {chromote}   JS websocket 11:09:59.10 send {"method":"update","data":{"bins":20}}
# {chromote}   JS websocket 11:09:59.11 recv {"progress":{"type":"binding",|truncated
# {chromote}   JS websocket 11:09:59.11 recv {"busy":"busy"}
# {chromote}   JS info      11:09:59.11 shinytest2; shiny:busy
# {chromote}   JS websocket 11:09:59.12 recv {"custom":{"showcase-src":{"srcref":|truncated
# {chromote}   JS websocket 11:09:59.14 recv {"recalculating":{"name":"distPlot",|truncated
# {chromote}   JS websocket 11:09:59.18 recv {"recalculating":{"name":"distPlot",|truncated
# {chromote}   JS websocket 11:09:59.19 recv {"busy":"idle"}
# {chromote}   JS info      11:09:59.19 shinytest2; shiny:idle
# {chromote}   JS websocket 11:09:59.21 recv {"errors":{},"values":{"distPlot":{|truncated
# {chromote}   JS info      11:09:59.21 shinytest2; shiny:value distPlot
# {shinytest2} R  info      11:09:59.21 Finished setting inputs. Timedout: FALSE
# {shinytest2} R  info      11:09:59.21 Getting all values
# {shiny}      R  error     ----------- Loading required package: shiny
# {shiny}      R  error     ----------- Running application in test mode.
# {shiny}      R  error     -----------
# {shiny}      R  error     ----------- Listening on http://127.0.0.1:1505
# {shiny}      R  error     ----------- SEND {"config":{"workerId":"","sessionId":|truncated
# {shiny}      R  error     ----------- RECV {"method":"init","data":{"bins":30,|truncated
# {shiny}      R  error     ----------- SEND {"custom":{"showcase-src":{"srcref"|truncated
# {shiny}      R  error     ----------- SEND {"busy":"busy"}
# {shiny}      R  error     ----------- SEND {"custom":{"showcase-src":{"srcref"|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"busy":"idle"}
# {shiny}      R  error     ----------- SEND {"errors":{},"values":{"distPlot":{|truncated
# {shiny}      R  error     ----------- RECV {"method":"update","data":{"bins":20}}
# {shiny}      R  error     ----------- SEND {"progress":{"type":"binding",|truncated
# {shiny}      R  error     ----------- SEND {"busy":"busy"}
# {shiny}      R  error     ----------- SEND {"custom":{"showcase-src":{"srcref":|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"recalculating":{"name":"distPlot",|truncated
# {shiny}      R  error     ----------- SEND {"busy":"idle"}
# {shiny}      R  error     ----------- SEND {"errors":{},"values":{"distPlot":{|truncated

}