Remote control a Shiny app running in a headless browser

This class starts a Shiny app in a new R session, along with a chromote headless browser that can be used to simulate user actions. This provides a full simulation of a Shiny app so that you can test user interactions with a live app.

Methods

Public methods

ShinyDriver2$expectAppshot()

ShinyDriver2$getDebugLog()

ShinyDriver2$enableDebugLogMessages()

ShinyDriver2$getAppDir()

ShinyDriver2$getAppFilename()

ShinyDriver2$isRmd()

ShinyDriver2$findElement()

ShinyDriver2$findElements()

ShinyDriver2$logEvent()

ShinyDriver2$getEventLog()

ShinyDriver2$executeScript()

ShinyDriver2$executeScriptCallback()

ShinyDriver2$expectUpdate()

ShinyDriver2$expectSnapshotJS()

ShinyDriver2$getAllValues()

ShinyDriver2$new()

ShinyDriver2$setInputs()

ShinyDriver2$uploadFile()

ShinyDriver2$takeScreenshot()

ShinyDriver2$expectDownload()

ShinyDriver2$stop()

ShinyDriver2$waitFor()

ShinyDriver2$waitForShiny()

ShinyDriver2$waitForValue()

ShinyDriver2$findWidget()

ShinyDriver2$getValue()

ShinyDriver2$setValue()

ShinyDriver2$click()

ShinyDriver2$listWidgets()

ShinyDriver2$checkUniqueWidgetNames()

ShinyDriver2$setWindowSize()

ShinyDriver2$getWindowSize()

ShinyDriver2$clone()

Method expectAppshot()

Usage

ShinyDriver2$expectAppshot(
  ...,
  name = NULL,
  items = NULL,
  screenshot = NULL,
  error = FALSE,
  cran = FALSE
)

Arguments

name

Name of a shiny widget.

Method getDebugLog()

Usage

ShinyDriver2$getDebugLog(type = c("all", debug_log_types()))

Method enableDebugLogMessages()

Usage

ShinyDriver2$enableDebugLogMessages(enable = TRUE)

Method getAppDir()

Usage

ShinyDriver2$getAppDir()

Method getAppFilename()

Usage

ShinyDriver2$getAppFilename()

Method isRmd()

Usage

ShinyDriver2$isRmd()

Method findElement()

Usage

ShinyDriver2$findElement(css = NULL)

Method findElements()

Usage

ShinyDriver2$findElements(css = NULL)

Method logEvent()

Usage

ShinyDriver2$logEvent(event, ...)

Method getEventLog()

Usage

ShinyDriver2$getEventLog()

Method executeScript()

Usage

ShinyDriver2$executeScript(script, ...)

Method executeScriptCallback()

Usage

ShinyDriver2$executeScriptCallback(script, ..., timeout_ = 15 * 1000)

Method expectUpdate()

Usage

ShinyDriver2$expectUpdate(
  output,
  ...,
  timeout = 3000,
  iotype = c("auto", "input", "output")
)

Arguments

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method expectSnapshotJS()

Usage

ShinyDriver2$expectSnapshotJS(script, ..., post_script = NULL, cran = FALSE)

Method getAllValues()

Usage

ShinyDriver2$getAllValues(input = TRUE, output = TRUE, export = TRUE)

Method new()

Usage

ShinyDriver2$new(
  path = ".",
  ...,
  load_timeout = NULL,
  screenshot = TRUE,
  check_names = TRUE,
  name = NULL,
  variant = getOption("shinytest2.variant", os_name_and_r_version()),
  debug = c("none", "all", debug_log_types()),
  seed = NULL,
  clean_logs = TRUE,
  shiny_args = list(),
  render_args = NULL,
  options = list()
)

Arguments

name

Name of a shiny widget.

Method setInputs()

Usage

ShinyDriver2$setInputs(
  ...,
  wait_ = TRUE,
  values_ = TRUE,
  timeout_ = 3000,
  allow_input_no_binding_ = FALSE,
  priority_ = c("input", "event")
)

Arguments

wait_

Wait until all reactive updates have completed?

Method uploadFile()

Usage

ShinyDriver2$uploadFile(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3000)

Arguments

wait_

Wait until all reactive updates have completed?

Method takeScreenshot()

Usage

ShinyDriver2$takeScreenshot(
  filename = NULL,
  ...,
  screenshot_args = list(),
  delay = 0,
  selector = "html",
  wait_ = TRUE
)

Arguments

wait_

Wait until all reactive updates have completed?

Method expectDownload()

Usage

ShinyDriver2$expectDownload(id, ..., name = NULL, cran = FALSE)

Arguments

name

Name of a shiny widget.

Method stop()

Usage

ShinyDriver2$stop()

Method waitFor()

Usage

ShinyDriver2$waitFor(expr, timeout = 3 * 1000, interval = 100)

Method waitForShiny()

Usage

ShinyDriver2$waitForShiny(timeout = 3 * 1000, interval = 100)

Method waitForValue()

Usage

ShinyDriver2$waitForValue(
  name,
  ignore = list(NULL, ""),
  iotype = c("input", "output", "export"),
  timeout = 10000,
  check_interval = 400
)

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method findWidget()

Usage

ShinyDriver2$findWidget(name, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method getValue()

Usage

ShinyDriver2$getValue(name, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method setValue()

Usage

ShinyDriver2$setValue(name, value, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method click()

Usage

ShinyDriver2$click(name, iotype = c("auto", "input", "output"))

Arguments

name

Name of a shiny widget.

iotype

Type of the Shiny widget. Usually shinytest2 finds the widgets by their name, so this is only needed if you use the same name for an input and output widget.

Method listWidgets()

Usage

ShinyDriver2$listWidgets()

Method checkUniqueWidgetNames()

Usage

ShinyDriver2$checkUniqueWidgetNames()

Method setWindowSize()

Usage

ShinyDriver2$setWindowSize(width, height)

Method getWindowSize()

Usage

ShinyDriver2$getWindowSize()

Method clone()

The objects of this class are cloneable with this method.

Usage

ShinyDriver2$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.